+++ src/index.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import { Provider } from 'react-redux';
+import { BrowserRouter } from 'react-router-dom';
 
 import store from '@src/store';
 import App from '@src/App';
@@ -9,7 +10,9 @@ import 'normalize.css';
 
 ReactDOM.render(
   <Provider store={store}>
-    <App />
+    <BrowserRouter forceRefresh>
+      <App />
+    </BrowserRouter>
   </Provider>,
   document.getElementById('root'),
 );
